+++ custom_components/prix_carburant/__init__.py
@@ -4,7 +4,12 @@ from datetime import timedelta
 import logging
 
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import ATTR_NAME, CONF_SCAN_INTERVAL, ATTR_LATITUDE, ATTR_LONGITUDE
+from homeassistant.const import (
+    ATTR_LATITUDE,
+    ATTR_LONGITUDE,
+    ATTR_NAME,
+    CONF_SCAN_INTERVAL,
+)
 from homeassistant.core import (
     HomeAssistant,
     ServiceCall,