+++ custom_components/prix_carburant/const.py
@@ -1,10 +1,11 @@
 """Constants for the Prix Carburant integration."""
+
 from typing import Final
 
 from homeassistant.const import Platform
 
 DOMAIN: Final = "prix_carburant"
-PLATFORMS: Final = [Platform.SENSOR, Platform.BUTTON]
+PLATFORMS: Final = [Platform.BUTTON, Platform.SENSOR]
 
 DEFAULT_NAME: Final = "Prix Carburant"
 DEFAULT_MAX_KM: Final = 15
@@ -15,8 +16,8 @@ ATTR_POSTAL_CODE = "postal_code"
 ATTR_BRAND = "brand"
 ATTR_CITY = "city"
 ATTR_DISTANCE = "distance"
-ATTR_FUEL_TYPE = "fuel_type"
 ATTR_FUELS = "fuels"
+ATTR_FUEL_TYPE = "fuel_type"
 ATTR_UPDATED_DATE = "updated_date"
 ATTR_DAYS_SINCE_LAST_UPDATE = "days_since_last_update"
 ATTR_PRICE = "price"